backend/requirements.txt
fastapi
uvicorn[standard]
websockets
sqlalchemy
psycopg2-binary
python-dotenv
openai-whisper
google-cloud-texttospeech
whisper-mic
fastmcp
google-generativeai
Jinja2
httpx
aiohttp
python-multipart
pydub
numpy
requests
starlette
imageio-ffmpeg